requirements.txt
django-allauth
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pymysql
dnspython